backend/function-audio-output/requirements.txt
functions-framework==3.*
flask==3.0.2
google-cloud-texttospeech==2.*
